Added Updates to Input component:
Updated select menus to now display the name of the nation and their currencies symbol rather than a three letter abberviation. This should make the user's experience with the site increase, as the update gives the selection menu a clear and easily readable interface.

Updated currency controller file on the server.
This update adds the use of a JSON file that has nation names and currency symbols, which is used when sending title data to the client.
This update is mainly used by the getTitle method where the server sends an array of objects containing a nation's currency code, symbol, and title.

Updated unorganized array of titles sent to the client from the getTitle method. The array is now sent in alphabetical order.